refactor: small improvements
