+++ app/components/Sidebar.tsx
@@ -5,12 +5,10 @@ import { usePaginatedQuery } from "convex/react";
 import { api } from "@/convex/_generated/api";
 import { useAuth } from "@workos-inc/authkit-nextjs/components";
 import { useRouter } from "next/navigation";
-import ChatItem from "./ChatItem";
 import { Button } from "@/components/ui/button";
-import { Plus, MessageSquare, PanelLeft } from "lucide-react";
+import { Plus, PanelLeft } from "lucide-react";
 import { useGlobalState } from "../contexts/GlobalState";
 import { useIsMobile } from "@/hooks/use-mobile";
-import Loading from "@/components/ui/loading";
 import {
   Sidebar,
   SidebarContent,
@@ -19,10 +17,11 @@ import {
   SidebarGroupContent,
   SidebarHeader,
 } from "@/components/ui/sidebar";
-import UserDropdownMenu from "./UserDropdownMenu";
+import SidebarHistory from "./SidebarHistory";
+import SidebarUserNav from "./SidebarUserNav";
 
 // Shared content components
-const ChatSidebarHeader: React.FC<{
+const SidebarHeaderContent: React.FC<{
   handleNewChat: () => void;
   handleCloseSidebar: () => void;
 }> = ({ handleNewChat, handleCloseSidebar }) => (
@@ -51,109 +50,7 @@ const ChatSidebarHeader: React.FC<{
   </div>
 );
 
-const ChatSidebarList: React.FC<{
-  chats: any[];
-  currentChatId: string | null;
-  handleNewChat: () => void;
-  paginationStatus?:
-    | "LoadingFirstPage"
-    | "CanLoadMore"
-    | "LoadingMore"
-    | "Exhausted";
-  loadMore?: (numItems: number) => void;
-  containerRef?: React.RefObject<HTMLDivElement | null>;
-}> = ({
-  chats,
-  currentChatId,
-  handleNewChat,
-  paginationStatus,
-  loadMore,
-  containerRef,
-}) => {
-  // Handle scroll to load more chats when scrolling to bottom
-  React.useEffect(() => {
-    const handleScroll = () => {
-      if (
-        !containerRef?.current ||
-        !loadMore ||
-        paginationStatus !== "CanLoadMore"
-      ) {
-        return;
-      }
-
-      const { scrollTop, scrollHeight, clientHeight } = containerRef.current;
-
-      // Check if we're near the bottom (within 100px)
-      if (scrollTop + clientHeight >= scrollHeight - 100) {
-        loadMore(28); // Load 28 more chats
-      }
-    };
-
-    const container = containerRef?.current;
-    if (container) {
-      container.addEventListener("scroll", handleScroll);
-      return () => container.removeEventListener("scroll", handleScroll);
-    }
-  }, [containerRef, loadMore, paginationStatus]);
-
-  if (paginationStatus === "LoadingFirstPage") {
-    // Loading state
-    return (
-      <div className="p-2">
-        <div className="space-y-3">
-          {[...Array(5)].map((_, i) => (
-            <div key={i} className="animate-pulse">
-              <div className="h-4 bg-sidebar-accent rounded w-3/4 mb-2"></div>
-              <div className="h-3 bg-sidebar-accent rounded w-1/2"></div>
-            </div>
-          ))}
-        </div>
-      </div>
-    );
-  }
-
-  if (!chats || chats.length === 0) {
-    // Empty state
-    return (
-      <div className="flex flex-col items-center justify-center h-full p-6 text-center">
-        <MessageSquare className="w-12 h-12 text-sidebar-accent-foreground mb-4" />
-        <h3 className="text-lg font-medium text-sidebar-foreground mb-2">
-          No chats yet
-        </h3>
-        <p className="text-sm text-sidebar-accent-foreground mb-4">
-          Start a conversation to see your chat history here
-        </p>
-        <Button onClick={handleNewChat} variant="default" size="sm">
-          <Plus className="w-4 h-4 mr-2" />
-          New Chat
-        </Button>
-      </div>
-    );
-  }
-
-  // Chat list with buttons (same for mobile and desktop)
-  return (
-    <div className="p-2 space-y-1">
-      {chats.map((chat: any) => (
-        <ChatItem
-          key={chat._id}
-          id={chat.id}
-          title={chat.title}
-          isActive={currentChatId === chat.id}
-        />
-      ))}
-
-      {/* Loading indicator when loading more */}
-      {paginationStatus === "LoadingMore" && (
-        <div className="flex justify-center py-2">
-          <Loading size={6} />
-        </div>
-      )}
-    </div>
-  );
-};
-
-const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
+const MainSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
   isMobileOverlay = false,
 }) => {
   const { user } = useAuth();
@@ -165,6 +62,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
     closeSidebar,
     setCurrentChatId,
     currentChatId,
+    initializeNewChat,
   } = useGlobalState();
 
   // Create ref for scroll container
@@ -188,9 +86,8 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
       setChatSidebarOpen(false);
     }
 
-    // Reset state first to ensure clean transition
-    resetChat();
-    setCurrentChatId(null);
+    // Initialize new chat state using global state function
+    initializeNewChat();
 
     // Navigate to homepage
     router.push("/");
@@ -206,7 +103,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
       <div className="flex flex-col h-full w-full bg-sidebar border-r">
         {/* Header */}
         <div className="p-2">
-          <ChatSidebarHeader
+          <SidebarHeaderContent
             handleNewChat={handleNewChat}
             handleCloseSidebar={handleCloseSidebar}
           />
@@ -215,7 +112,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
         {/* Chat List */}
         <div className="flex-1 overflow-hidden">
           <div ref={scrollContainerRef} className="h-full overflow-y-auto">
-            <ChatSidebarList
+            <SidebarHistory
               chats={paginatedChats.results || []}
               currentChatId={currentChatId}
               handleNewChat={handleNewChat}
@@ -227,9 +124,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
         </div>
 
         {/* Footer */}
-        <div className="border-t border-sidebar-border">
-          <UserDropdownMenu />
-        </div>
+        <SidebarUserNav />
       </div>
     );
   }
@@ -237,7 +132,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
   return (
     <Sidebar side="left" className={`${isMobile ? "w-full" : "w-72"}`}>
       <SidebarHeader>
-        <ChatSidebarHeader
+        <SidebarHeaderContent
           handleNewChat={handleNewChat}
           handleCloseSidebar={handleCloseSidebar}
         />
@@ -247,7 +142,7 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
         <SidebarGroup>
           <SidebarGroupContent>
             <div ref={scrollContainerRef} className="h-full overflow-y-auto">
-              <ChatSidebarList
+              <SidebarHistory
                 chats={paginatedChats.results || []}
                 currentChatId={currentChatId}
                 handleNewChat={handleNewChat}
@@ -261,12 +156,10 @@ const ChatSidebar: React.FC<{ isMobileOverlay?: boolean }> = ({
       </SidebarContent>
 
       <SidebarFooter>
-        <div className="border-t border-sidebar-border">
-          <UserDropdownMenu />
-        </div>{" "}
+        <SidebarUserNav />
       </SidebarFooter>
     </Sidebar>
   );
 };
 
-export default ChatSidebar;
+export default MainSidebar;
