--- app/components/UserDropdownMenu.tsx
@@ -1,99 +1,0 @@
-"use client";
-
-import React from "react";
-import { useAuth } from "@workos-inc/authkit-nextjs/components";
-import { LogOut, User } from "lucide-react";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-
-const UserDropdownMenu = () => {
-  const { user } = useAuth();
-
-  if (!user) return null;
-
-  const handleSignOut = async () => {
-    window.location.href = "/logout";
-  };
-
-  const getUserInitials = () => {
-    const firstName = user.firstName?.charAt(0)?.toUpperCase() || "";
-    const lastName = user.lastName?.charAt(0)?.toUpperCase() || "";
-    if (firstName && lastName) {
-      return firstName + lastName;
-    }
-    if (firstName) {
-      return firstName;
-    }
-    if (lastName) {
-      return lastName;
-    }
-    return user.email?.charAt(0)?.toUpperCase() || "U";
-  };
-
-  const getDisplayName = () => {
-    if (user.firstName && user.lastName) {
-      return `${user.firstName} ${user.lastName}`;
-    }
-    return user.firstName || user.lastName || "User";
-  };
-
-  return (
-    <DropdownMenu>
-      <DropdownMenuTrigger asChild>
-        <div className="flex items-center gap-3 p-3 cursor-pointer hover:bg-sidebar-accent/50 rounded-md transition-colors">
-          <Avatar className="h-8 w-8">
-            <AvatarImage
-              src={user.profilePictureUrl || undefined}
-              alt={getDisplayName()}
-            />
-            <AvatarFallback className="text-xs">
-              {getUserInitials()}
-            </AvatarFallback>
-          </Avatar>
-          <div className="flex-1 min-w-0">
-            <div className="text-sm font-medium text-sidebar-foreground truncate">
-              {getDisplayName()}
-            </div>
-            <div className="text-xs text-sidebar-accent-foreground truncate">
-              {user.email}
-            </div>
-          </div>
-        </div>
-      </DropdownMenuTrigger>
-
-      <DropdownMenuContent
-        className="w-56"
-        align="end"
-        side="top"
-        sideOffset={8}
-      >
-        <DropdownMenuLabel className="font-normal">
-          <div className="flex flex-col space-y-1">
-            <p className="text-sm font-medium leading-none">
-              {getDisplayName()}
-            </p>
-            <p className="text-xs leading-none text-muted-foreground">
-              {user.email}
-            </p>
-          </div>
-        </DropdownMenuLabel>
-
-        <DropdownMenuSeparator />
-
-        <DropdownMenuItem onClick={handleSignOut} variant="destructive">
-          <LogOut className="mr-2 h-4 w-4" />
-          <span>Log out</span>
-        </DropdownMenuItem>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-};
-
-export default UserDropdownMenu;
